Refactor code for build deployment
